requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.1
pydantic-settings==2.8.1
replicate==1.0.7
pillow==11.1.0
rembg==2.0.61
onnxruntime==1.22.1
redis==5.0.8